+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - `symbol_unsorted` option for the `table` component.
 - `caption` assign for the `table` component.
-- `style` assign for the `:col` slot of the `table` component.
+- `col_style` assign for the `:col` slot of the `table` component.
 
 ## [0.13.0] - 2021-11-14